Learners Permit
So I kind of forgot how to use any of these logics when I started and got incredibly stumped on the first tables and sub questions, which felt necessary before moving on. I'm going to keep working on this, while giving myself grace, and respect if this will be the submission that counts. Thank you for being so empathetic with us.